Refactor backend names to pillow and wand
This uses absolute_import and refactors the backend names to pillow &
wand. It changes tests to use these names and also modified the test
runner to not pass any parameters concerning the image backends since the
defaults are enough for tests. It also fixes a minor bug that was occured
when the image file had already been read (multiple renditions in a single
page)
