+++ wagtail/wagtailimages/backends/pillow.py
